+++ src/tile-atlas.hpp
@@ -29,7 +29,7 @@ struct tile_atlas final
 
 private:
     static std::unique_ptr<const texcoords[]> make_texcoords_array(Vector2ui pixel_size, Vector2ub tile_count);
-    static texcoords make_texcoords(Vector2ui pixel_size, Vector2ub tile_count, std::uint8_t i);
+    static texcoords make_texcoords(Vector2ui pixel_size, Vector2ub tile_count, std::size_t i);
 
     std::unique_ptr<const texcoords[]> texcoords_;
     GL::Texture2D tex_;
